requirements.txt
pandas
python-dotenv
streamlit

langchain
langchain-community
langchain-openai

faiss-cpu

openai
tqdm
requests
